Correct version of CIGRE MV benchmar network
It turned out that I initially uploaded a version of the MATPOWER case had not been converted directly from the pandapower case as previously said. In the new MATPOWER case there is no branch between buses 14 and 8. Another difference is that the new MATPOWER case does not include any gencost data.