Change storage Dict key "disharge_rate" -> "discharge_rating"
Got error message that the key "disharge_rate" did not exist in the storage dictionary.  This error message appeared both for tests of the storage model and of the demand flexibility model. Made changes(/fixes?) to be able to run cases without errors. However, discharge_rate is also used as a variable within some storage.jl functions, so these variables are not renamed. If some of the changed dictionary keys represented a variable distinct from the key "discharge_rating", this commit may have introduced errors...