Update case6_flex.m with DC branches from "Italy" case
Copy set-up of AC and DC branches (existing and candidates) from case6_all_candidates.m to case6.m to make sure the branches and their numbers are consistent across different tests. Need to redo plotting to illustrate the flow to bus 5 (now it is the two existing DC branches that connects bus 5 ("Sardinia") to buses 1 and 3 ("mainland Italy"), in addition to a potential DC branch from node 6 ("Sicilia").

On test_flex_demand