Test forcing congestion for transformer to feeder 1
Forcing congestion for transformer to feeder 1 and plotting demand flexibility for all (flexible) loads on feeder 1, so needed to fix plotting that previously only considered a single flexible load. Also needed to fix bug due to local/global scope ambiguity. But still there seems to be a bug in the plotting of the baseline/flexible load demand in feeder 1 compared to the power flowing over the transformer, and there seems to be approx. 1 MW load going missing on bus 1(?!)

On test_flex_demand