Updated deprecated syntax of `CSV.read()`
When `CSV.read()` is used without a sink function as the second argument, an `ArgumentError` or a `Warning` are produced, depending on the version of `CSV` package.
This commit updates each call to `CSV.read()` by adding a `DataFrame` as a sink function. The result is exactly the same as before (if no sink function was specified, `DataFrame` was used by default).