Fixed a bug in `constraint_shift_duration()`
`start_grace` nw id must be greater than or equal to the id that the same network had in the first optimization period (i.e. the first id in network_ids variable in the new notation). This id can be different from 1 in problems that use scenarios or subnetworks.

This commit restores the approach used until commit `39a10a58ffd30ca229c3c8651bb4ede92e3449c6` in problem functions and constraint templates, i.e. pass to the constraint template a collection of all the relevant network ids.
An alternative implementation would be to pass only the id of the first nw.