Candidate branch activation variables have the same value in all periods
Fixes a bug introduced in commit `c2c827f3cfbac7b7c7fac2a0806979fe85651c1f` as a result of which the candidate branch binary activation variables could have different values in consecutive optimization periods when using the distribution model.
This commit restores the previous behavior, i.e. that the activation variables are bound to assume the same value in consecutive periods.
Differently from before, however, the involved constraint is now placed together with the other intertemporal constraints concerning storage and flexible loads, so that it can be applied correctly even when using a combined transmission and distribution model, where a hardcoded `n > 1` would not produce the desired result.