Fixed too limiting bounds on power of `ne_branch`es added in parallel
`ne_branch`es intended to be added in parallel to existing `branch`es are modeled through replacement with the parallel equivalent.
The parameters that must be specified in input data are the parameters of the candidate branch only: those of the parallel equivalent are computed later in the involved constraints: `constraint_voltage_magnitude_difference()` and `constraint_thermal_limit_from()`.
By that time, `variable_ne_branch_power()` has already computed the bounds on the power variables of `ne_branch`es, using the rated power of the candidate branch only instead of the parallel equivalent: these bounds are too limiting.

This commit avoids the problem by not generating bounds on the power variables of `ne_branches`.

Another approach would be to compute the parallel equivalent at preprocessing stage, i.e. making a function that takes the input data structure, substitutes the parameters of the parallel equivalent to the parameters of the candidate branch and returns the updated data structure.