Benders’ decomposition procedure (draft)
In PowerModels, PowerModelsACDC and FlexPlan there are some network components whose existence depends on a binary activation variable. In particular, there is one binary variable for each component AND each optimization period. However, there is also a constraint that requires that all the binary variables relating to the same component have the same value.

Benders' procedure relies on a decomposition in which the binary variables are repeatedly fixed at different values. In this procedure it may be beneficial to reduce the number of binary variables (and the number of associated constraints) by removing the dependance on the optimization period.

This commit removes the dependance of the binary variables on the optimization period (just for `flex_tnep` problem for the moment, but other problems can be adapted with minimal effort) by generating in the JuMP model only the variable of the first period and associating it with the variables of all periods. This approach has the following drawbacks:

- requires a custom implementation for each function that generates binary variables;
- requires that the for loop that generates the variables in the problem functions iterates over the networks in ascending order.

Current limitations: the procedure only supports the `flex_tnep` problem; in particular, only one scenario is considered.