Add unit tests: flexible demand model for CIGRE MV benchmark network
Including some simple unit tests before starting to clean up the code. Now load shifting is enabled in the case, so that it should reproduce the situation shown in Fig. 55 in D1.2 where flexibility is activated and no new line is built; previously it reproduced Fig. 54 where candidate branch 1 was built but no flexibility was activated (which seemed like a more boring situation to test). But these tests should rather be moved to a separate file where there are now user-specified input parameter as in this runscript.

On fix_66_clean_up_demand_flexibility