Remove ref. to parameters cost_shift_up_tot_max and cost_shift_down_tot_max
The code for moving flexible demand data from the load_extra to the load dictionary is looking for parameters cost_shift_up_tot_max and cost_shift_down_tot_max that do not exist. This does not cause any crashes but is rather unnecessary.

On fix_66_clean_up_demand_flexibility