Fix paths for unit tests
Changing paths for unit tests to expect the code to be run from the test directory when being run as tests. Also making sure absolute paths are used (and correctly so) in loading the demand data for the CIGRE MV benchmark case for flexibility, following the suggestion by matteorossini.

On fix_66_clean_up_demand_flexibility