Reorganise code for cut generation, improve time reporting
Now, when an optimality cut is generated, the iteration over the integer variables uses the same logic as `get_var_values()`; therefore, it is no longer necessary to manually specify the names of the integer variables, so `int_vars` is removed from procedure parameters.
In addition, optimality cut generation is moved to the beginning of the iterations.

Time reporting is improved and now the plot produced by `test/scripts/test_benders_decomposition.jl` reports also the time spent in each iteration outside main and secondary problems.