Speedup Benders' procedure by building the solution only at the end
The last iteration may not be the one that provides the best solution.
The previous approach to deal with this was to build the solution in each improving iteration, but it had the drawback of requiring a consderable amount of time (up to 60% of the entire procedure!).
The current approach is:
- In each improving iteration, store only the value of the integer variables;
- At the end of the procedure:
  1. If the last iteration is not the best one, then fix the integer variables to the values provided by the best iteration and solve the secondary problems again;
  2. Build the solution.