Add the modern implementation of Benders decomposition
The modern implementation uses solver-independent callbacks (lazy constraints) to solve secondary problems whenever an optimal integer solution of the main problem is found.

Stopping at a certain tolerance with respect to the optimal solution is not directly supported in the procedure, but can be achieved by passing an appropriate parameter to the optimizer (`CPXPARAM_MIP_Tolerances_MIPGap` for CPLEX) as done in `test/scripts/test_benders_decomposition.jl`.

Stopping the procedure when the iteration limit is reached is supported.
If the iteration limit is reached, the callback returns without solving secondary problems. The solver of the main problem might invoke the callback a few times again, but should terminate soon after.
The procedure returns the best solution found before hitting the iteration limit.