New system to manage multinetwork dimensions
`add_dimension!()` is introduced to attach to a single-network data structure the information needed to handle multiple _dimensions_, like hours, years, scenarios, subnetworks and so on.
For each dimension, its _cardinality_ must be specified either as a number or through a Dict of _properties_ (for example, a probability for each scenario). Additional _metadata_, referring to the dimension as a whole, can also be specified.

A bunch of new functions is introduced to handle dimensions in problem functions:
- `nw_ids()` to get the sorted nw ids of an `AbstractPowerModel`, optionally filtered by the value of one or more dimensions;
- `dim()` and `dim_meta()` to get the properties and the metadata, respectively, of a dimension;
- `is_first_nw()` and `is_last_nw()` to tell whether a give nw id is the first/last along a dimension;
- `first_nw()`, `last_nw()`, `prev_nw()` and `next_nw()` to get a related nw id moving along a specified dimension.

The function `multinetwork_data()` is updated to exploit the data structure made available by `add_dimension!()`.

Combined T&D models are not yet supported.