Update problem functions
Change way of adding intertemporal constraints: instead of grouping by scenario (which is not extensible to multiple years or new dimensions), check if the current nw is related to the first/last hour and add the constraint accordingly.
This new approach requires a rewriting of some flexibility constraints (`constraint_shift_duration` and related ones) to pass them the id of the first nw along the `hour` dimension.