Define/redefine variables for investment candidates
Now for each investment candidate there are two functions that generate variables: one whose name ends in `_indicator`, that generates a variable aimed at tracking the activation of the candidate, and one whose name ends in `_investment`, that generates a variable aimed at tracking the investment decision related to the candidate.

For each investment candidate, the same pair of variables is shared among all hours and scenarios.