Update objective functions
Auxiliary functions that calculate investment costs now track the investment binary variable instead of the indicator variable: the cost is charged when the investment decision is taken.

- `objective_stoch_flex`: updated.
- `objective_min_cost_storage`: updated.
- `objective_min_cost_flex`: updated transmission function; distribution function need to be updated after subnetworks are fixed.
- `objective_reliability`: not updated (requires partial rewriting).

For now, the residual value of investments at the end of the planning horizon is not taken into account, so all unit tests should pass after updating input files.