Update `test/test_flex_demand.jl`
Some tests in this file have failed for weeks due to changes in cost scaling.
Some costs in `test/data/case6_flex.m` are updated to produce the same investment decisions in the optimal solution. Furthermore, the cost for the time shift of the flex load is set to a strictly positive value in order to obtain stability of the results across different CI configurations.
Some values in `test/test_flex_demand.jl` are changed to match the updated test case result.