Add keys from time series even if they are not found in template nw
In addition to modifying the values of keys that are already present in `template_nw`, data from `time_series` can now also be used to add values for keys that are not present in `template_nw`.
Thanks to this, it is no longer necessary to create a `template_nw` with dummy values which will then have to be replaced by those taken from `time_series`.