Add curtailment of non-dispatchable generators
Non-dispatchable generators can now be curtailed.
The curtailed power is implemented through an "expression" that replaces the symbol `:pgcurt` with the difference between the reference power of the generator and the power actually produced.
A term is added in the objective function to account for the curtailment cost, coefficient of which is specified as `cost_curt` in Matpower files.
The curtailed power is reported in the solution dictionary (also for dispatchable generators, for which it is zero).