Use `_PM.` instead of `PowerModels.` in code
When importing PowerModels with `import PowerModels; const _PM = PowerModels` both the identifiers can be used.
However, when importing PowerModels with `import PowerModels as _PM` (Julia 1.6+) only `_PM` is available.
Therefore, it is best to only use `_PM` in the code. The documentation on the other hand can continue to use `PowerModels` in full, which is probably clearer to the reader.