Delete unmaintained code
This commit deletes:
- code that depends on JuliaDB;
- code related to reliability analysis;
- code that is no longer used after previous deletions.

JuliaDB is basically unmaintained (latest release is one year and a half ago; multiple blocking issues refer to <https://github.com/JuliaData/JuliaDB.jl/pull/389> which is stale) and cannot be used together with the other dependencies due to unsatisfiable version requirements. Substituting JuliaDB with DataFrames would entail a complete rewrite in most cases.

The code related to reliability analysis has not kept up with the updates made in the repository in the last year (such as adding dimensions, new problem functions, investment constraints specified only once a year rather than for all hours...).

Any code deleted here can of course be restored when someone is willing to update it.