Allow functions that load test cases to replicate constant data
By default, functions that load test cases do not replicate constant data (i.e., variables from different networks may be bound to the same value in memory). This is good for speed and usually harmless, as data dictionaries are typically soon passed to functions that build optimization problems.
However, sometime it may be necessary to apply additional transformations to data dictionaries. In such cases, the property that data from different networks are indpendent of each other may be required.