Improve numerical stability
In the decoupling procedure, some optimization problems are solved in sequence. Upper/lower bounds are set on some variables of a problem based on values provided by the optimal solution of a problem solved earlier. Due to feasibility tolerance, for some of those variables in the second problem it can happen that the upper bound is slightly (around 1e-16 or so) lower than the lower bound, making the second problem infeasible. In that case, the new bound is increased or decreased to be equal to the existing bound.