Rotate the octagon used in `BFA8PowerModels`'s branch thermal limit
The octagon is still inscribed in a circle with a radius of `rate_a`, but with respect to the previous configuration, in which the octagon had 4 edeges parallel to the main axes, now is rotated by pi/8, so that the maximum absolute values for active and reactive power can be reached. Previously, they were limited to ~ 0.92 * rate_a.